+++ src/hooks/query/financialReports.js
@@ -304,12 +304,11 @@ export function useCustomerBalanceSummaryReport(query, props) {
     },
     {
       select: (res) => ({
-        columns: res.data.columns,
         query: res.data.query,
-        tableRows: res.data.table.rows,
+        table: res.data.table,
       }),
       defaultData: {
-        tableRows: [],
+        table: {},
         query: {},
       },
       ...props,
@@ -334,12 +333,11 @@ export function useVendorsBalanceSummaryReport(query, props) {
 
     {
       select: (res) => ({
-        columns: res.data.columns,
         query: res.data.query,
-        tableRows: res.data.table.data,
+        table: res.data.table,
       }),
       defaultData: {
-        tableRows: [],
+        table: {},
         query: {},
       },
       ...props,